model_scripts/requirements.txt
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
spacy==3.6.1
requests==2.31.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
crypt4gh
cryptography
